Updates the tests with user-defined sizes
Since the interfaces for IsotropicFunction and AnisotropicFunction has
changed, these tests needed to be updated.

All clients using the above two classes have to change the getValue
member to take in and apf::Vector3 of coordinates, instead of the
vertex, and return the corresponding user defined sizes (or frames in
case of anisotropic size-fields).

The main reason for this change is to be able to used these for higher
order representation of the size fields as well.